Simplify player composable
Move all the state, loading and unloading to the activity so the composable can focus in the player itself
