+++ app/src/main/java/cc/wordview/app/ui/activities/player/PlayerActivity.kt
@@ -22,23 +22,47 @@ import android.view.WindowManager
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import cc.wordview.app.SongViewModel
 import cc.wordview.app.extensions.setOrientationSensorLandscape
 import cc.wordview.app.extractor.DownloaderImpl
+import cc.wordview.app.extractor.VideoStream
+import cc.wordview.app.misc.AppSettings
 import cc.wordview.app.misc.ImageCacheManager
+import cc.wordview.app.ui.activities.player.composables.ErrorScreen
 import cc.wordview.app.ui.activities.player.composables.Player
+import cc.wordview.app.ui.activities.player.viewmodel.PlayerState
+import cc.wordview.app.ui.activities.player.viewmodel.PlayerViewModel
+import cc.wordview.app.ui.components.CircularProgressIndicator
 import cc.wordview.app.ui.components.OneTimeEffect
 import cc.wordview.app.ui.theme.WordViewTheme
+import cc.wordview.gengolex.Language
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import me.zhanghai.compose.preference.LocalPreferenceFlow
 import me.zhanghai.compose.preference.ProvidePreferenceLocals
 import org.schabi.newpipe.extractor.NewPipe
+import org.schabi.newpipe.extractor.exceptions.ExtractionException
 import timber.log.Timber
 
 @AndroidEntryPoint
 class PlayerActivity : ComponentActivity() {
+    private val viewModel: PlayerViewModel by viewModels()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -46,26 +70,81 @@ class PlayerActivity : ComponentActivity() {
         setupWindowInsets()
         enableEdgeToEdge()
         setContent {
-            val context = LocalContext.current
+            ProvidePreferenceLocals {
+                val state by viewModel.playerState.collectAsStateWithLifecycle()
+                val videoId by SongViewModel.videoId.collectAsStateWithLifecycle()
+                val videoStream by SongViewModel.videoStream.collectAsStateWithLifecycle()
+                val errorMessage by viewModel.errorMessage.collectAsStateWithLifecycle()
+                val statusCode by viewModel.statusCode.collectAsStateWithLifecycle()
 
-            OneTimeEffect {
-                // If no trees were planted, we can assume that
-                // this activity was started separately
-                if (Timber.treeCount == 0) {
-                    DownloaderImpl.init(null)
-                    NewPipe.init(DownloaderImpl.getInstance())
+                val preferences by LocalPreferenceFlow.current.collectAsStateWithLifecycle()
+                val langTag = AppSettings.language.get()
 
-                    Timber.plant(Timber.DebugTree())
-                    ImageCacheManager.init(context)
+                val context = LocalContext.current
+
+                OneTimeEffect {
+                    // If no trees were planted, we can assume that
+                    // this activity was started separately
+                    if (Timber.treeCount == 0) {
+                        DownloaderImpl.init(null)
+                        NewPipe.init(DownloaderImpl.getInstance())
+
+                        Timber.plant(Timber.DebugTree())
+                        ImageCacheManager.init(context)
+                    }
                 }
-            }
 
-            WordViewTheme {
-                ProvidePreferenceLocals {
-                    Player()
+                fun start() {
+                    val lang = Language.byTag(langTag)
+
+                    Timber.i("Chosen language is ${lang.name.lowercase()}")
+
+                    CoroutineScope(Dispatchers.IO).launch {
+                        try {
+                            SongViewModel.videoStream.value.init(videoId, context)
+
+                            viewModel.initAudio(videoStream.getStreamURL(), context)
+                            viewModel.getLyrics(preferences, context, videoId, lang, videoStream)
+                        } catch (e: ExtractionException) {
+                            Timber.e(e)
+                            viewModel.setErrorMessage(e.message.toString())
+                            viewModel.setPlayerState(PlayerState.ERROR)
+                        }
+                    }
+                }
+
+                OneTimeEffect { start() }
+
+                WordViewTheme {
+                    Scaffold { innerPadding ->
+                        when (state) {
+                            PlayerState.READY -> Player(viewModel, innerPadding)
+
+                            PlayerState.ERROR -> ErrorScreen(errorMessage, {
+                                Timber.d("Refreshing player")
+                                viewModel.setPlayerState(PlayerState.LOADING)
+                                start()
+                            }, statusCode)
+
+                            PlayerState.LOADING -> Box(
+                                Modifier.fillMaxSize(),
+                                contentAlignment = Alignment.Center
+                            ) {
+                                CircularProgressIndicator(64.dp)
+                            }
+                        }
+                    }
                 }
             }
         }
+    }
+
+    override fun onDestroy() {
+        if (isFinishing) {
+            SongViewModel.setVideoStream(VideoStream())
+        }
+
+        super.onDestroy()
     }
 
     private fun setupWindowInsets() {
@@ -75,7 +154,8 @@ class PlayerActivity : ComponentActivity() {
         controller.hide(WindowInsetsCompat.Type.systemBars())
         controller.hide(WindowInsetsCompat.Type.displayCutout())
 
-        controller.systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
+        controller.systemBarsBehavior =
+            WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
 
         window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
     }
